pyproject.toml
[project]
name = "heidi-todoist"
version = "0.1.0"
description = "Azure function to complete Todoist tasks for Heidi"
authors = [
    {name = "Tom Boone",email = "tbc@tomboone.com"}
]
readme = "README.md"
license = "MIT"
license-files = [
    "LICENSE",
]
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "azure-functions (>=1.23.0,<2.0.0)",
    "todoist-api-python (>=3.1.0,<4.0.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)",
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.13.1"
pytest = "^8.0.0"
pytest-cov = "^7.0.0"
pytest-mock = "^3.10.0"
responses = "^0.25.0"
pre-commit = "^4.3.0"
bandit = "^1.8.6"
mypy = "^1.18.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


